// src/store/action-types.ts
export enum ActionType {
  LOGIN = 'LOGIN',
  LOGOUT = 'LOGOUT',
  FETCH_TASKS = 'FETCH_TASKS',
  SET_LOADING = 'SET_LOADING',
  UPDATE_SORT_FIELD = 'UPDATE_SORT_FIELD',
  UPDATE_SORT_DIRECTION = 'UPDATE_SORT_DIRECTION',
  UPDATE_PAGE = 'UPDATE_PAGE',
  ADD_MESSAGE = 'ADD_MESSAGE',
  REMOVE_MESSAGE = 'REMOVE_MESSAGE',
}